+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.1.5-alpha
+
+* Add Snowflake data stream commands
+* Update Snowflake integration creation to require assigning a RAI engine
+
 ## v0.1.4-alpha
 
 * Add Snowflake integration commands